Ship zero-JS-framework static site; React stays build-time only
The blank page reproduced on multiple devices with no JS errors while
the served HTML verifiably contained full content — the only remaining
moving part was the React runtime replacing the DOM at boot. Remove it:

- React renders the site at build time only (prerender pipeline keeps
  components, data files, and tests unchanged)
- Browser gets static HTML plus a 1.4KB vanilla script (theme toggle,
  back-to-top); was a 216KB React bundle
- Scroll-reveal removed; entry animations remain CSS-only and safe
- Verified: page complete with JS disabled (11 papers, 20 news items);
  with JS: zero errors, root intact, toggle works
