+++ index.html
@@ -105,7 +105,7 @@
         }, 2000);
       })();
     </script>
-    <script type="module" src="/src/main.tsx"></script>
+    <script type="module" src="/src/site.ts"></script>
     <script>
       // Visit /?debug=1 to overlay render diagnostics on the page.
       if (location.search.indexOf('debug') !== -1) {
