+++ src/components/BackToTop.tsx
@@ -1,23 +1,6 @@
-import { useEffect, useState } from 'react';
-
 export default function BackToTop() {
-  const [visible, setVisible] = useState(false);
-
-  useEffect(() => {
-    const onScroll = () => setVisible(window.scrollY > 600);
-    onScroll();
-    window.addEventListener('scroll', onScroll, { passive: true });
-    return () => window.removeEventListener('scroll', onScroll);
-  }, []);
-
   return (
-    <button
-      type="button"
-      className={`back-to-top${visible ? ' visible' : ''}`}
-      onClick={() => window.scrollTo({ top: 0, behavior: 'smooth' })}
-      aria-label="Back to top"
-      tabIndex={visible ? 0 : -1}
-    >
+    <button type="button" className="back-to-top" aria-label="Back to top">
       <span aria-hidden="true">↑</span>
     </button>
   );
