+++ src/components/Footer.tsx
@@ -1,10 +1,6 @@
-import { useScrollReveal } from '../hooks/useScrollReveal.ts';
-
 export default function Footer() {
-  const footerRef = useScrollReveal<HTMLElement>();
-
   return (
-    <footer className="site-footer" ref={footerRef}>
+    <footer className="site-footer">
       <div className="footer-line">&copy; {new Date().getFullYear()} Sharath Chandra Raparthy</div>
       <div className="footer-meta">
         Built with React + Vite ·{' '}
