--- src/hooks/useScrollReveal.ts
@@ -1,40 +1,0 @@
-import { useEffect, useRef } from 'react';
-
-/**
- * Adds the `scroll-reveal` class to the referenced element and flips it to
- * `revealed` once it scrolls into view (mirrors the original inline script).
- */
-export function useScrollReveal<T extends HTMLElement>() {
-  const ref = useRef<T>(null);
-
-  useEffect(() => {
-    const el = ref.current;
-    if (!el) return;
-
-    el.classList.add('scroll-reveal');
-
-    const observer = new IntersectionObserver(
-      (entries) => {
-        entries.forEach((entry) => {
-          if (entry.isIntersecting) {
-            entry.target.classList.add('revealed');
-            observer.unobserve(entry.target);
-          }
-        });
-      },
-      { threshold: 0.1, rootMargin: '0px 0px -40px 0px' },
-    );
-
-    observer.observe(el);
-
-    // Failsafe: never leave content hidden if the observer misbehaves.
-    const failsafe = setTimeout(() => el.classList.add('revealed'), 1500);
-
-    return () => {
-      observer.disconnect();
-      clearTimeout(failsafe);
-    };
-  }, []);
-
-  return ref;
-}
